declare phony rules for special make targets
In terms of Make, a phony target is simply a target that is always
out-of-date, so whenever you ask 'make <phony_target>', it will run,
independent from the state of the file system. Some common make
targets that are often phony are: all, install, clean.

https://www.gnu.org/software/make/manual/html_node/Phony-Targets.html

If a rule has no prerequisites or recipe, and the target of the rule
is a nonexistent file, then make imagines this target to have been
updated whenever its rule is run. This implies that all targets
depending on this one will always have their recipe run. The target
simz80.c is such kind of rule to ensure it will generated by simz80.pl
independently fom the state of the file system.

https://www.gnu.org/software/make/manual/html_node/Force-Targets.html

Signed-off-by: Stephan Linz <linz@li-pro.net>