avoid warnings of ignored return value
The void cast of result value is the standard convention to say
"this result is unused intentionally" and universally recognized,
except by gcc because someone decided 20 years ago that there
should be no way to ignore a warn_unused_result. Many compilers
will optimize away this void cast, as it is unnecessary, and then
an unused variable error will still be produced. GCC will still
warn even when you do this. See this discussion for more info:

    http://gcc.gnu.org/bugzilla/show_bug.cgi?id=25509

Signed-off-by: Stephan Linz <linz@li-pro.net>